Removes NewConversationComponent
